Workaround for sampling in CombinedPrFunction.
There seems to be an issue with the findIndex method which is used by invcdf,
which ultimatly is used by the sample method. The workaround is that
sample(Decimal) iterates over the exact values and buckets until it has
found the right value or bucket. At least for exact values the rounding
issues seem to be fixed by using numberic_limits<Decimal>::epsilon().